% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcovcases.R
\name{get_country_cases_data}
\alias{get_country_cases_data}
\title{Get specific country's COVID-19 cases and deaths from WHO COVID-19 Global Data.}
\usage{
get_country_cases_data(country_name = "The United Kingdom", days = 90)
}
\arguments{
\item{country_name}{target country for data visualisation, default "The United Kingdom".}

\item{days}{date duration, default 90.}
}
\description{
Get specific country's COVID-19 cases and deaths from WHO COVID-19 Global Data.
}
