% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcovcases.R
\name{plot_country_cumulative_deaths}
\alias{plot_country_cumulative_deaths}
\title{Visualise specific country's COVID-19 cumulative deaths from WHO COVID-19 Global Data.}
\usage{
plot_country_cumulative_deaths(
  country_name = "The United Kingdom",
  days = 90,
  axis_x_txt_size = 8,
  line_size = 2,
  line_colour = "#3EC70B"
)
}
\arguments{
\item{country_name}{target country of data visualisation, default "The United Kingdom".}

\item{days}{date duration, default 90.}

\item{axis_x_txt_size}{size of X axis text, default 8.}

\item{line_size}{size of line, default 2.0,}

\item{line_colour}{colour of line, default "#3EC70B".}
}
\description{
Visualise specific country's COVID-19 cumulative deaths from WHO COVID-19 Global Data.
}
