% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcovcases.R
\name{plot_country_new_cases_deaths}
\alias{plot_country_new_cases_deaths}
\title{Visualise specific country's relationships between cases and deaths using connected scatter plot.}
\usage{
plot_country_new_cases_deaths(
  country_name = "The United Kingdom",
  days = 90,
  axis_x_txt_size = 8,
  segment_size = 0.02,
  segment_colour = "gray",
  point_size = 2,
  point_colour = "#3EC70B",
  label_size = 2,
  label_colour = "#3EC70B"
)
}
\arguments{
\item{country_name}{target country of data visualisation, default "The United Kingdom".}

\item{days}{date duration, default 90.}

\item{axis_x_txt_size}{size of X axis text, default 8.}

\item{segment_size}{segment size of connected scatter plot, default 0.02.}

\item{segment_colour}{segment colour of connected scatter plot, default "gray".}

\item{point_size}{point size of connected scatter plot, default 2.0.}

\item{point_colour}{point colour of connected scatter plot, default "#3EC70B".}

\item{label_size}{label size of connected scatter plot, default 2.0.}

\item{label_colour}{label colour of connected scatter plot, default "#3EC70B".}
}
\description{
Visualise specific country's relationships between cases and deaths using connected scatter plot.
}
