% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcovcases.R
\name{plot_country_new_deaths}
\alias{plot_country_new_deaths}
\title{Visualise specific country's COVID-19 new deaths from WHO COVID-19 Global Data.}
\usage{
plot_country_new_deaths(
  country_name = "The United Kingdom",
  days = 90,
  axis_x_txt_size = 8,
  line_size_deaths = 1,
  line_size_average = 2,
  line_colour_deaths = "gray",
  line_colour_average = "#3EC70B"
)
}
\arguments{
\item{country_name}{target country of data visualisation, default "The United Kingdom".}

\item{days}{date duration, default 90.}

\item{axis_x_txt_size}{size of X axis text, default 8.}

\item{line_size_deaths}{size of deaths line, default 1.0.}

\item{line_size_average}{size of 7 days average line, default 2.0.}

\item{line_colour_deaths}{colour of deaths line, default "gray".}

\item{line_colour_average}{colour of 7 days average line, "#3EC70B".}
}
\description{
Visualise specific country's COVID-19 new deaths from WHO COVID-19 Global Data.
}
